parseDouble doesn't exist but praseFloat does.
git-svn-id: file:///Users/gauthierm/svnrepo/xml-rpc-ajax/trunk@2792 94490950-a86c-4493-9ccb-530f72933e47

+++ www/javascript/XML_RPC_Types.js
@@ -168,7 +168,7 @@ XML_RPC_Double.prototype.marshall = function()
 XML_RPC_Double.unmarshall = function(double_node)
 {
 	var value = double_node.firstChild.nodeValue;
-	var double_value = parseDouble(value);
+	var double_value = parseFloat(value);
 	return double_value;
 }
